refactor(plugin-workflow-mailer): change plugin name and locales
