+++ packages/plugins/@nocobase/plugin-workflow-mailer/.npmignore
