+++ packages/plugins/@nocobase/plugin-workflow-mailer/LICENSE
