+++ packages/plugins/@nocobase/plugin-workflow-mailer/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@nocobase/plugin-workflow-smtp-mailer",
-  "displayName": "Workflow: send email",
-  "displayName.zh-CN": "工作流：发送邮件",
-  "description": "Send email by calling SMTP service in workflow.",
-  "description.zh-CN": "可用于在工作流中向通过 SMTP 服务发送邮件。",
+  "name": "@nocobase/plugin-workflow-mailer",
+  "displayName": "Workflow: mailer node",
+  "displayName.zh-CN": "工作流：邮件发送节点",
+  "description": "Send email in workflow.",
+  "description.zh-CN": "可用于在工作流中发送电子邮件。",
   "version": "1.0.0-alpha.17",
   "license": "AGPL-3.0",
   "main": "./dist/server/index.js",
