+++ packages/plugins/@nocobase/plugin-workflow-mailer/src/locale/index.ts
