+++ packages/plugins/@nocobase/plugin-workflow-mailer/src/locale/zh-CN.json
@@ -1,0 +1,19 @@
+{
+  "Mailer": "邮件发送",
+  "Send email. You can use the variables in the upstream nodes as receivers, subject and content of the email.": "发送邮件。可以使用上游节点中的变量作为邮件的接收者、主题和内容。",
+  "SMTP host": "SMTP 服务主机名",
+  "Port": "端口",
+  "Secure": "安全模式",
+  "User": "用户名",
+  "Password": "密码",
+  "From": "发件人",
+  "To": "收件人",
+  "CC": "抄送",
+  "BCC": "暗送",
+  "Add email address": "新增邮件地址",
+  "Subject": "主题",
+  "Content": "内容",
+  "Content type": "内容格式",
+  "Plain text": "纯文本",
+  "Ignore failed sending and continue workflow": "忽略失败的发送并继续工作流"
+}
