--- packages/plugins/@nocobase/plugin-workflow-smtp-mailer/src/locale/zh-CN.json
@@ -1,19 +1,0 @@
-{
-  "SMTP mailer": "SMTP 邮件发送器",
-  "Send email by calling SMTP service. You can use the variables in the upstream nodes as receivers, subject and content of the email.": "通过调用 SMTP 服务发送邮件。可以使用上游节点中的变量作为邮件的接收者、主题和内容。",
-  "Service host": "邮件服务主机名",
-  "Port": "端口",
-  "Secure": "安全模式",
-  "User": "用户名",
-  "Password": "密码",
-  "From": "发件人",
-  "To": "收件人",
-  "CC": "抄送",
-  "BCC": "暗送",
-  "Add email address": "新增邮件地址",
-  "Subject": "主题",
-  "Content": "内容",
-  "Content type": "内容格式",
-  "Plain text": "纯文本",
-  "Ignore failed sending and continue workflow": "忽略失败的请求并继续工作流"
-}
